+++ src/components/container/EhrNewsFeed/EhrNewsFeed.stories.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { Layout } from "../../../presentational";
+import { Layout } from "../../presentational";
 import EhrNewsFeed, { EhrNewsFeedProps } from "./EhrNewsFeed";
 
-export default { title: "EhrNewsFeed/Container/EhrNewsFeed", component: EhrNewsFeed, parameters: { layout: 'fullscreen' } };
+export default { title: "Container/EhrNewsFeed", component: EhrNewsFeed, parameters: { layout: 'fullscreen' } };
 let render = (args: EhrNewsFeedProps) => <Layout colorScheme="auto">
     <EhrNewsFeed {...args} />
 </Layout>
