+++ src/components/container/EhrNewsFeedEventDetail/EhrNewsFeedEventDetail.stories.tsx
@@ -1,22 +1,15 @@
 import React from "react";
-import { Layout } from "../../../presentational";
-import EventDetail, { EventDetailProps } from "./EventDetail";
+import { Layout } from "../../presentational";
+import EhrNewsFeedEventDetail, { EhrNewsFeedEventDetailProps } from "./EhrNewsFeedEventDetail";
 
-export default { title: "EhrNewsFeed/Container/EventDetail", component: EventDetail, parameters: { layout: 'fullscreen' } };
-let render = (args: EventDetailProps) => <Layout colorScheme="auto">
-    <EventDetail {...args} />
+export default { title: "EhrNewsFeed/Container/EventDetail", component: EhrNewsFeedEventDetail, parameters: { layout: 'fullscreen' } };
+let render = (args: EhrNewsFeedEventDetailProps) => <Layout colorScheme="auto">
+    <EhrNewsFeedEventDetail {...args} />
 </Layout>
 
 export const ProcedureGroup = {
     args: {
         previewState: "ProcedureGroup"
-    },
-    render: render
-};
-
-export const Report = {
-    args: {
-        previewState: "Report"
     },
     render: render
 };
