--- src/components/ehr-news-feed/presentational/LabReportDetail/LabReportDetail.css
@@ -1,35 +1,0 @@
-.mdhui-lab-report-detail .mdhui-lab-report-detail-normal-range {
-    font-size: .7em;
-    color: var(--mdhui-text-color-2);
-}
-
-.mdhui-lab-report-detail .mdhui-lab-report-detail-comment {
-    white-space: pre-wrap;
-    font-family: monospace;
-    font-size: .7em;
-    padding: 16px;
-    border-top: solid 1px var(--mdhui-border-color-1);
-}
-
-.mdhui-lab-report-detail .mdhui-lab-report-detail-value {
-    display: flex;
-    align-items: center;
-    gap: var(--mdhui-padding-sm);
-}
-
-.mdhui-lab-report-detail .mdhui-lab-report-detail-acuity {
-    display: block;
-    font-size: .7em;
-    color: #FFF;
-    padding-left: var(--mdhui-padding-xxs);
-    padding-right: var(--mdhui-padding-xxs);
-    border-radius: 4px;
-}
-
-.mdhui-lab-report-detail .mdhui-lab-report-detail-acuity-low {
-    background: var(--mdhui-color-primary);
-}
-
-.mdhui-lab-report-detail .mdhui-lab-report-detail-acuity-high {
-    background: var(--mdhui-color-danger);
-}
