--- src/components/ehr-news-feed/presentational/NewsFeedListItem/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./NewsFeedListItem";
