--- src/components/ehr-news-feed/presentational/index.ts
@@ -1,1 +1,0 @@
-export { default as EhrNewsFeedProcedureGroupListItem } from "./NewsFeedListItem"
